cpp6 cleaned up. Ready to push

+++ 5_cpp/ex00/Bureaucrat.hpp
@@ -17,7 +17,7 @@ class Bureaucrat
 {
 	private:
 		const std::string	_Name;
-		short				_Rank;
+		int					_Rank;
 		void				changeRank(int);
 	protected:
